requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
python-dotenv==1.0.0

# File processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf==3.17.0

# Data processing
pandas==2.1.4
numpy==1.26.2
tqdm==4.66.1

# Development tools (optional)
pytest==7.4.3
black==23.12.1
flake8==6.1.0

# Additional utilities
requests==2.31.0
pydantic==2.5.2